Move junixsocket and Unix SocketChannel implementations in separate modules
